feat: Added script which syncs repositories as files
Why not git submodules? Because we want to snapshot the files in time
and not rely on some upstream. Whatever is in the template now should
work.

Secondly we might want multiple versions of the same template which
would make submodules a bit confusing (multiple versions of the same
submodule?).

Lastly, the wasmopticon will most likely be a submodule it self to other
repositories, at least the integration tests and nested git submodules
are the spawn of all evil.